Add files via upload
Complete program used in the portable LAMP device, with two individual timers for each chamber, detention of use of external power or a power bank, and regulation of voltage delivered to the heating element depending on temperature and power source. There is also personalization to help the user better perform the assays by giving all the necessary information through a display screen, LEDs, and buzzer sounds.